feat: verify user & get certain user's transactions

+++ controller/transactionController.js
@@ -25,6 +25,11 @@ export class TransactionController {
         next(new HttpError(404, `Transaction with ID ${id} not found.`));
       }
 
+      if (transaction.userId !== res.locals.user.userId) {
+        next(new HttpError(403, "Forbidden: You do not own this transaction."));
+        return;
+      }
+
       return res.status(200).send(transaction);
     } catch (error) {
       next(error);
@@ -33,7 +38,7 @@ export class TransactionController {
 
   getAllTransactionsByUserId = async (req, res, next) => {
     try {
-      const userId = req.params.userId;
+      const userId = res.locals.user.userId; // get from the payload
       const transactions =
         await this.transactionRepository.getAllTransactionsByUserId(userId);
       return res.status(200).send(transactions);
@@ -44,8 +49,8 @@ export class TransactionController {
 
   createTransaction = async (req, res, next) => {
     try {
-      const { userId, accountId, amount, transactionDate, description } =
-        req.body;
+      const userId = res.locals.user.userId;
+      const { accountId, amount, transactionDate, description } = req.body;
       const newTransaction = await this.transactionRepository.createTransaction(
         userId,
         accountId,
@@ -64,6 +69,21 @@ export class TransactionController {
     try {
       const id = req.params.id;
       const updates = req.body;
+
+      const transaction = await this.transactionRepository.getTransactionById(
+        id
+      );
+
+      if (!transaction) {
+        next(new HttpError(404, `Transaction with ID ${id} not found.`));
+        return;
+      }
+
+      if (transaction.userId !== res.locals.user.userId) {
+        next(new HttpError(403, "Forbidden: You do not own this transaction."));
+        return;
+      }
+
       const updatedTransaction =
         await this.transactionRepository.updateTransactionById(id, updates);
 
@@ -80,10 +100,16 @@ export class TransactionController {
 
       if (!deleted) {
         next(new HttpError(404, `Transaction with ID ${id} not found.`));
+        return;
+      }
+
+      if (deleted.userId !== res.locals.user.userId) {
+        next(new HttpError(403, "Forbidden: You do not own this transaction."));
+        return;
       }
 
       await this.transactionRepository.deleteTransactionById(id);
-      return res.status(204).send();
+      return res.status(204).send(`Transaction with ID ${id} deleted.`);
     } catch (error) {
       next(error);
     }
